% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_wid_data}
\alias{get_wid_data}
\title{Download WID data}
\usage{
get_wid_data(indicators = "sptinc")
}
\arguments{
\item{indicators}{Indiciator to download, defaults to "sptinc". See  for additional indicators and more.}
}
\value{
A tibble containing the income data.
}
\description{
Download WID data
}
