% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/income-package.R
\docType{package}
\name{income}
\alias{income}
\alias{income-package}
\title{\pkg{income}: A package for plotting the Lorenz-Curve.}
\description{
The income package is implemented via \pkg{drake}. The Workflow is the following:
\enumerate{
\item Execute \code{make_project()}
\item Visualise the \pkg{drake} plan with \code{show_project}
\item Plot Lorenz-Curves with \code{plot_lorenz()}. Country ID's can be seen with
\code{country_list()}
}
}
