% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drake_plan.R
\name{make_project}
\alias{make_project}
\title{"Make" the project with.}
\usage{
make_project(jobs = 1)
}
\arguments{
\item{jobs}{Number of jobs for parallel processing with.}
}
\description{
"Make" the project with.
}
