% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_lorenz}
\alias{plot_lorenz}
\title{Plot Lorenz Curve for a specified country}
\usage{
plot_lorenz(ctry = "GB")
}
\arguments{
\item{wid_data}{A tibble imported via.}

\item{country}{A string giving the country.}
}
\value{
A plot of the Lorenz Curve
}
\description{
Plot Lorenz Curve for a specified country
}
